third_party/WebKit/LayoutTests/platform/win/http/tests/security/mixedContent/websocket/insecure-websocket-in-secure-page-expected.txt

CONSOLE ERROR: line 12: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-throw-on-construction.html' was loaded over HTTPS, but attempted to connect to the insecure WebSocket endpoint 'ws://example.test:8880/echo'. This request has been blocked; this endpoint must be available over WSS.
This is a testharness.js-based test.
PASS Test that when a HTTPS window that creates an insecure WebSocket connection is opened, the mixed content callback is triggered because the main frame in the window is HTTPS but is trying connecting to an insecure WebSocket server. 
Harness: the test ran to completion.
